Split the util directory
