+++ backend/app.py
@@ -11,7 +11,7 @@ from api.result.route import result_api_bp
 from api.swagger.route import swagger_api_bp
 from api.system.route import system_api_bp
 from api.test_case.route import test_case_api_bp
-from utils.sandbox_util import execute_queueing_task_when_exist_empty_box
+from api.judge.sandbox_util import execute_queueing_task_when_exist_empty_box
 
 from flask import Flask
